+++ tests/Specification/Logical/NotSpecificationTest.php
@@ -1,0 +1,36 @@
+<?php
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHP\Specification\Logical;
+
+use ExtendsSoftware\ExaPHP\Specification\SpecificationInterface;
+use PHPUnit\Framework\TestCase;
+
+class NotSpecificationTest extends TestCase
+{
+    /**
+     * Is satisfied.
+     *
+     * Test that specification is satisfied when inner specification is not satisfied.
+     *
+     * @covers \ExtendsSoftware\ExaPHP\Specification\Logical\NotSpecification::__construct()
+     * @covers \ExtendsSoftware\ExaPHP\Specification\Logical\NotSpecification::isSatisfied()
+     */
+    public function testIsSatisfied(): void
+    {
+        $specification = $this->createMock(SpecificationInterface::class);
+        $specification
+            ->expects($this->exactly(2))
+            ->method('isSatisfied')
+            ->with('foo')
+            ->willReturnOnConsecutiveCalls(true, false);
+
+        /**
+         * @param SpecificationInterface $specification
+         */
+        $notSpecification = new NotSpecification($specification);
+
+        $this->assertFalse($notSpecification->isSatisfied('foo'));
+        $this->assertTrue($notSpecification->isSatisfied('foo'));
+    }
+}