+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Added ProxyValidator to proxy value and context to inner validator.
 - Added object support to FlattenerInterface::flatten() method.
 - Added NoTagsValidator to validate string without HTML-tags.
+- Added SpecificationInterface and implementations to use for specification pattern.
 
 ### Changed